Order System
Developed an e-commerce website where a customer can view, edit and
order parts in a catalogue. And, once ordered the parts, it will be
redirected to a warehouse interface where warehouse admin can look up in
inventory for the ordered parts, create shipping labels, and then ships
the product to customer. The project was developed using HTML, PHP
(Server scripting language) and MySQL for storing the customer, as well
as to maintain parts inventory database.